+++ lib/errors/index.ts
@@ -1,5 +1,7 @@
 export { AlreadyStoppedError } from './AlreadyStoppedError';
+export { LowAudioLevelWarning } from './LowAudioLevelWarning';
 export { DiagnosticError } from './DiagnosticError';
+export { DiagnosticWarning } from './DiagnosticWarning';
 export { InvalidStateError } from './InvalidStateError';
 export { PromiseTimedOutError } from './PromiseTimedOutError';
 export { UnsupportedError } from './UnsupportedError';